Align main menu for lf design
