+++ frontend/src/assets/scss/index.scss
@@ -33,6 +33,7 @@
 @import 'form/checkbox';
 @import 'form/switch';
 @import 'tooltip';
+@import 'layout/index';
 
 @import '../../../config/styles';
 @import '../../ui-kit';
