+++ frontend/src/assets/scss/layout/index.scss
@@ -1,0 +1,1 @@
+@import "menu";
