+++ frontend/src/modules/layout/components/layout.vue
@@ -2,7 +2,7 @@
   <el-container v-if="tenant" class="flex-col">
     <el-container style="height: calc(100vh - 60px); margin-top: 60px;">
       <!-- App menu -->
-      <app-lf-menu />
+      <lf-menu />
 
       <el-container class="bg-white rounded-tl-2xl">
         <el-main id="main-page-wrapper" class="relative">
@@ -14,25 +14,18 @@
   </el-container>
 </template>
 
-<script>
-import AppLfMenu from '@/modules/lf/layout/components/lf-menu.vue';
+<script setup lang="ts">
 import AppLfBanners from '@/modules/lf/layout/components/lf-banners.vue';
 import { useAuthStore } from '@/modules/auth/store/auth.store';
 import { storeToRefs } from 'pinia';
+import LfMenu from '@/modules/layout/components/menu/menu.vue';
 
+const authStore = useAuthStore();
+const { tenant } = storeToRefs(authStore);
+</script>
+
+<script lang="ts">
 export default {
   name: 'AppLayout',
-
-  components: {
-    AppLfMenu,
-    AppLfBanners,
-  },
-
-  setup() {
-    const authStore = useAuthStore();
-    const { token, tenant } = storeToRefs(authStore);
-
-    return { tenant, token };
-  },
 };
 </script>
